backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
python-multipart==0.0.6
aiofiles==23.2.1
python-dotenv==1.0.0
supabase==2.0.2
requests==2.31.0
langchain==0.1.0
langchain-openai==0.0.5
openai>=1.10.0
python-dateutil==2.8.2